サイズが２の時の例の追加と概要の書き方の修正
